Split and rename modules and add restart button
